[#28] chore : manager 관련 타입 types 폴더로 이동
