+++ catch-study/src/pages/ManagementInfoPage/ManagementInfo.tsx
@@ -3,43 +3,11 @@ import AddressForm from '../../components/AddressForm/AddreesFrom';
 import FeeForm from '../../components/FeeForm/FeeForm';
 import RoomForm from '../../components/RoomForm/RoomForm';
 import Topbar from '../../components/Topbar/Topbar';
-
-export interface AddressTypes {
-  city: string;
-  country: string;
-  town: string;
-  etc: string;
-}
-
-export interface RoomInfoTypes {
-  name: string;
-  capacity: number;
-}
-
-export interface UsageFeeTypes {
-  hours: number;
-  price: number;
-}
-
-interface FormDataTypes {
-  cafeName: string;
-  address: AddressTypes;
-  openingHours: string;
-  closedHours: string;
-  closedDay: string;
-  seats: number;
-  roomInfo: [
-    {
-      cancelAvailableTime: number;
-      rooms: RoomInfoTypes[];
-    },
-  ];
-  usageFee: UsageFeeTypes[];
-  titleImage: string;
-  multipleImages: string[];
-  seatChartImage: string;
-  cafePhone: string;
-}
+import {
+  FormDataTypes,
+  RoomInfoTypes,
+  UsageFeeTypes,
+} from '../../types/manage';
 
 const ManagementInfo: React.FC = () => {
   const [roomInfos, setRoomInfos] = useState<RoomInfoTypes[]>([]);
